requirements.txt
streamlit
langchain-groq
langchain-community
langchain-core
PyPDF2
python-dotenv
